+++ src/applications/vaos/tests/new-appointment/components/TypeOfFacilityPage.unit.spec.jsx
@@ -1,77 +1,32 @@
 import React from 'react';
-import { TypeOfFacilityPage } from '../../../new-appointment/components/TypeOfFacilityPage';
-import { mockParentSites, mockSupportedFacilities } from '../../mocks/helpers';
-import { getFacilityMock, getParentSiteMock } from '../../mocks/v0';
+import { expect } from 'chai';
+import { mockFetch, resetFetch } from 'platform/testing/unit/helpers';
 import {
   createTestStore,
-  renderWithStoreAndRouter,
   setTypeOfCare,
+  renderWithStoreAndRouter,
 } from '../../mocks/setup';
-import { mockFetch, resetFetch } from 'platform/testing/unit/helpers';
+
+import { fireEvent } from '@testing-library/dom';
+import TypeOfFacilityPage from '../../../new-appointment/components/TypeOfFacilityPage';
 
 const initialState = {
   featureToggles: {
     vaOnlineSchedulingVSPAppointmentNew: false,
     vaOnlineSchedulingDirect: true,
-    // eslint-disable-next-line camelcase
-    show_new_schedule_view_appointments_page: true,
   },
   user: {
     profile: {
-      facilities: [
-        { facilityId: '983', isCerner: false },
-        { facilityId: '984', isCerner: false },
-      ],
+      facilities: [{ facilityId: '983', isCerner: false }],
     },
   },
 };
 
-const parentSite983 = {
-  id: '983',
-  attributes: {
-    ...getParentSiteMock().attributes,
-    institutionCode: '983',
-    authoritativeName: 'Some VA facility',
-    rootStationCode: '983',
-    parentStationCode: '983',
-  },
-};
-const parentSite984 = {
-  id: '984',
-  attributes: {
-    ...getParentSiteMock().attributes,
-    institutionCode: '984',
-    authoritativeName: 'Some other VA facility',
-    rootStationCode: '984',
-    parentStationCode: '984',
-  },
-};
-
-describe('VAOS integration: VA type of facility page with a multi-site user', () => {
+describe('VAOS integration: VA facility page with a single-site user', () => {
   beforeEach(() => mockFetch());
   afterEach(() => resetFetch());
 
-  it('should show form with required questions for both sites and facilities', async () => {
-    mockParentSites(['983', '984'], [parentSite983, parentSite984]);
-    const facility = {
-      id: '984',
-      attributes: {
-        ...getFacilityMock().attributes,
-        institutionCode: '984',
-        city: 'Bozeman',
-        stateAbbrev: 'MT',
-        authoritativeName: 'Bozeman VA medical center',
-        rootStationCode: '984',
-        parentStationCode: '984',
-        requestSupported: true,
-      },
-    };
-    mockSupportedFacilities({
-      siteId: '984',
-      parentId: '984',
-      typeOfCareId: '323',
-      data: [facility],
-    });
+  it('should show form with single required facility question', async () => {
     const store = createTestStore(initialState);
     await setTypeOfCare(store, /primary care/i);
 
@@ -79,110 +34,56 @@ describe('VAOS integration: VA type of facility page with a multi-site user', ()
       store,
     });
 
-    await screen.findByText(/Choose where you want to receive your care/i);
-    // expect(screen.getByLabelText(/some other va facility/i)).to.have.attribute(
-    //   'value',
-    //   'var984',
-    // );
+    // Check for page title
+    expect(screen.baseElement).to.contain.text(
+      'Choose where you want to receive your care',
+    );
+
+    // Check for page subtitle
+    expect(screen.baseElement).to.contain.text(
+      'You’re eligible to see either a VA provider or community care provider for this type of care.(*Required)',
+    );
+
+    // Check for option 1 radio
+    expect(
+      screen.getByRole('radio', {
+        name: /va medical center or clinic go to a va medical center or clinic for this appointment/i,
+      }),
+    ).to.have.attribute('value', 'vamc');
+
+    // Check for option 2 radio
+    expect(
+      screen.getByRole('radio', {
+        name: /community care facility go to a community care facility near your home/i,
+      }),
+    ).to.have.attribute('value', 'communityCare');
+
+    // Click on the continue button without an option selected
+    fireEvent.click(screen.getByRole('button', { name: /continue »/i }));
+
+    // Check for alert to be shown
+    await expect(screen.getByRole('alert')).to.contain.text(
+      'Please provide a response',
+    );
+
+    // Click on option 2 radio
+    fireEvent.click(
+      screen.getByRole('radio', {
+        name: /community care facility go to a community care facility near your home/i,
+      }),
+    );
+
+    // Click on the continue button with an option selected
+    fireEvent.click(screen.getByRole('button', { name: /continue »/i }));
+
+    // Check for next page to be navigated to
+    await expect(screen.baseElement).to.contain.text(
+      'Choose a day and time for your appointment',
+    );
+
+    // const radio = getByLabelText('First')
+    // fireEvent.change(radio, { target: { value: "second" } });
+    // expect(radio.value).toBe('second')
+    // fireEvent.click(screen.getByRole('radio', { name: /community care facility go to a community care facility near your home/i }));
   });
 });
-
-// describe('VAOS <TypeOfFacilityPage>', () => {
-//   it('should render', () => {
-//     const openFormPage = sinon.spy();
-//     const updateFormData = sinon.spy();
-//
-//     const form = mount(
-//       <TypeOfFacilityPage
-//         openFormPage={openFormPage}
-//         updateFormData={updateFormData}
-//         data={{}}
-//       />,
-//     );
-//
-//     expect(form.find('input').length).to.equal(2);
-//     form.unmount();
-//   });
-//
-//   it('should not submit empty form', () => {
-//     const openFormPage = sinon.spy();
-//     const history = {
-//       push: sinon.spy(),
-//     };
-//
-//     const form = mount(
-//       <TypeOfFacilityPage
-//         openFormPage={openFormPage}
-//         history={history}
-//         data={{}}
-//       />,
-//     );
-//
-//     form.find('form').simulate('submit');
-//
-//     expect(form.find('.usa-input-error').length).to.equal(1);
-//     expect(history.push.called).to.be.false;
-//     form.unmount();
-//   });
-//
-//   it('should call updateFormData after change', () => {
-//     const openFormPage = sinon.spy();
-//     const updateFormData = sinon.spy();
-//     const history = {
-//       push: sinon.spy(),
-//     };
-//
-//     const form = mount(
-//       <TypeOfFacilityPage
-//         openFormPage={openFormPage}
-//         updateFormData={updateFormData}
-//         history={history}
-//         data={{}}
-//       />,
-//     );
-//
-//     selectRadio(form, 'root_facilityType', FACILITY_TYPES.COMMUNITY_CARE);
-//
-//     expect(updateFormData.firstCall.args[2].facilityType).to.equal(
-//       FACILITY_TYPES.COMMUNITY_CARE,
-//     );
-//     form.unmount();
-//   });
-//
-//   it('should submit with valid data', () => {
-//     const openFormPage = sinon.spy();
-//     const routeToNextAppointmentPage = sinon.spy();
-//
-//     const form = mount(
-//       <TypeOfFacilityPage
-//         openFormPage={openFormPage}
-//         routeToNextAppointmentPage={routeToNextAppointmentPage}
-//         data={{ facilityType: FACILITY_TYPES.COMMUNITY_CARE }}
-//       />,
-//     );
-//
-//     form.find('form').simulate('submit');
-//
-//     expect(form.find('.usa-input-error').length).to.equal(0);
-//     expect(routeToNextAppointmentPage.called).to.be.true;
-//     form.unmount();
-//   });
-//
-//   it('document title should match h1 text', () => {
-//     const openFormPage = sinon.spy();
-//     const updateFormData = sinon.spy();
-//     const pageTitle = 'Choose where you want to receive your care';
-//
-//     const form = mount(
-//       <TypeOfFacilityPage
-//         openFormPage={openFormPage}
-//         updateFormData={updateFormData}
-//         data={{}}
-//       />,
-//     );
-//
-//     expect(form.find('h1').text()).to.equal(pageTitle);
-//     expect(document.title).contain(pageTitle);
-//     form.unmount();
-//   });
-// });
